% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_custom_functions.R
\name{hdp_extract_comp_multi}
\alias{hdp_extract_comp_multi}
\title{Extracts components from multiple MCMC chains}
\usage{
hdp_extract_comp_multi(chains, cos.merge = 0.9, sig.level = 0.95,
  redo = TRUE)
}
\description{
Extracts components from multiple MCMC chains
}

