% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_custom_functions.R
\name{hdp_extract_components}
\alias{hdp_extract_components}
\title{Extracts components from an hdp sample chain or multi-chain}
\usage{
hdp_extract_components(x, sig.level = 0.95, cos.merge = 0.9, redo = TRUE)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti object}

\item{sig.level}{Significance level to use when calculating HPD credible intervals (default 0.95)}

\item{cos.merge}{Merge components with cosine similarity above this threshold (default 0.90)}

\item{redo}{Logical. If true - constituent chains with previously calculated components will be re-calculated. Only used for hdpSampleMulti.}
}
\description{
Extracts components from an hdp sample chain or multi-chain
}

