% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_custom_functions.R
\name{merge_elems}
\alias{merge_elems}
\title{function to merge elements of a list according to new (numeric) names
ans will be sorted by numeric element names (is this still needed?)}
\usage{
merge_elems(lst, labels)
}
\description{
function to merge elements of a list according to new (numeric) names
ans will be sorted by numeric element names (is this still needed?)
}

