% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_custom_functions.R
\name{hdp_extract_comp_single}
\alias{hdp_extract_comp_single}
\title{Extracts components from single MCMC chain}
\usage{
hdp_extract_comp_single(chain, sig.level = 0.95, cos.merge = 0.9)
}
\description{
Extracts components from single MCMC chain
}

