% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_custom_functions.R
\name{merge_cols}
\alias{merge_cols}
\title{function to merge cols of a matrix according to new (numeric) column labels
ans will be sorted by numeric column name}
\usage{
merge_cols(mx, labels)
}
\arguments{
\item{mx}{The matrix to merge}

\item{labels}{The (new) set of labels for the matrix}
}
\description{
function to merge cols of a matrix according to new (numeric) column labels
ans will be sorted by numeric column name
}

